+++ dbversioning/repositoryconf.py
@@ -375,6 +375,32 @@ class RepositoryConf(object):
         return True
 
     @staticmethod
+    def remove_repo_version(repo_name: str, version_nums: str):
+        conf = RepositoryConf._get_repo_dict()
+        with open(RepositoryConf.config_file_name()):
+            repos = conf[REPOSITORIES_PROP]
+            rp = [r for r in repos if r[NAME_PROP] == repo_name]
+
+            if rp[0][ENVS_PROP]:
+                for k, v in rp[0][ENVS_PROP].items():
+                    if v == version_nums:
+                        del rp[0][ENVS_PROP][k]
+                        break
+
+            out_str = json.dumps(
+                    conf,
+                    indent=4,
+                    sort_keys=True,
+                    separators=(",", ": "),
+                    ensure_ascii=True,
+            )
+
+            with open(RepositoryConf.config_file_name(), "w") as outfile:
+                outfile.write(out_str)
+
+        return True
+
+    @staticmethod
     def get_repo_env(repo_name: str, env: str):
         conf = RepositoryConf._get_repo_dict()